requirements.txt
# Application requirements (with MySQL 8 support)
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
python-dotenv==1.0.0
Werkzeug==2.3.7
SQLAlchemy==2.0.21
click==8.1.7
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
blinker==1.6.3

# MySQL Database support
PyMySQL==1.1.0
mysqlclient==2.2.0
bcrypt==4.0.1

# Development tools
Flask-DebugToolbar==0.13.1
pytest==7.4.2
pytest-flask==1.2.0
black==23.9.1
flake8==6.1.0
isort==5.12.0
ipython==8.15.0
requests==2.31.0
Flask-Bcrypt==1.0.1
